refactor: FeedItem 리팩토링
